+++ mss/tests/test_gnu_linux.py
@@ -9,7 +9,7 @@ import platform
 
 import mss
 import pytest
-from mss.base import MSSMixin
+from mss.base import MSSBase
 from mss.exception import ScreenShotError
 
 
@@ -32,7 +32,7 @@ def test_factory_systems(monkeypatch):
     # GNU/Linux
     monkeypatch.setattr(platform, "system", lambda: "LINUX")
     with mss.mss() as sct:
-        assert isinstance(sct, MSSMixin)
+        assert isinstance(sct, MSSBase)
     monkeypatch.undo()
 
     # macOS