+++ mss/tests/test_implementation.py
@@ -6,59 +6,43 @@ Source: https://github.com/BoboTiG/python-mss
 import os
 import os.path
 import platform
-import sys
 
+import pytest
 import mss
 import mss.tools
-from mss.base import MSSMixin
+from mss.base import MSSBase
 from mss.exception import ScreenShotError
 from mss.screenshot import ScreenShot
 
-import pytest
 
-
-PY3 = sys.version[0] > "2"
-
-
-class MSS0(MSSMixin):
+class MSS0(MSSBase):
     """ Nothing implemented. """
 
     pass
 
 
-class MSS1(MSSMixin):
-    """ Emulate no monitors. """
+class MSS1(MSSBase):
+    """ Only `grab()` implemented. """
+
+    def grab(self, monitor):
+        pass
+
+
+class MSS2(MSSBase):
+    """ Only `monitor` implemented. """
 
     @property
     def monitors(self):
         return []
 
 
-class MSS2(MSSMixin):
-    """ Emulate one monitor. """
-
-    @property
-    def monitors(self):
-        return [{"top": 0, "left": 0, "width": 10, "height": 10}]
+@pytest.mark.parametrize("cls", [MSS0, MSS1, MSS2])
+def test_incomplete_class(cls):
+    with pytest.raises(TypeError):
+        cls()
 
 
-def test_incomplete_class():
-    # `monitors` property not implemented
-    with pytest.raises(NotImplementedError):
-        for filename in MSS0().save():
-            assert os.path.isfile(filename)
-
-    # `monitors` property is empty
-    with pytest.raises(ScreenShotError):
-        for filename in MSS1().save():
-            assert os.path.isfile(filename)
-
-    # `grab()` not implemented
-    sct = MSS2()
-    with pytest.raises(NotImplementedError):
-        sct.grab(sct.monitors[0])
-
-    # Bad monitor
+def test_bad_monitor(sct):
     with pytest.raises(ScreenShotError):
         sct.grab(sct.shot(mon=222))
 
@@ -79,7 +63,7 @@ def test_repr(sct, pixel_ratio):
 def test_factory(monkeypatch):
     # Current system
     with mss.mss() as sct:
-        assert isinstance(sct, MSSMixin)
+        assert isinstance(sct, MSSBase)
 
     # Unknown
     monkeypatch.setattr(platform, "system", lambda: "Chuck Norris")
@@ -87,10 +71,7 @@ def test_factory(monkeypatch):
         mss.mss()
     monkeypatch.undo()
 
-    if not PY3:
-        error = exc.value[0]
-    else:
-        error = exc.value.args[0]
+    error = exc.value.args[0]
     assert error == "System 'chuck norris' not (yet?) implemented."